Moved Views to new directory structure
