+++ src/Routing.ts
@@ -1,6 +1,6 @@
 import { SearchRouteProps, searchBase } from "./Search/Routing";
 import { SubjectsRouteProps, subjectBase } from "./subject/Routing";
-import { ViewRouteProps, viewBase } from "./views/Routing";
+import { ViewRouteProps, viewBase } from "./model/Views/Routing";
 
 import { RouteComponentProps } from "react-router";
