+++ src/views/components/ViewsNav/ViewsNav.tsx
@@ -13,8 +13,8 @@ import { useDispatch, useSelector } from "react-redux";
 import { RouteComponentProps } from "react-router";
 import { State } from "../../../Reducer";
 import { commandBarStyle } from "../../../AppCommandBar/AppCommandBar";
-import { createView } from "../../model/CreateView";
-import { gotoView } from "../../Routing";
+import { createView } from "../../../model/Views/CreateView";
+import { gotoView } from "../../../model/Views/Routing";
 import { subjectBase } from "../../../subject/Routing";
 import { useGetCreateButtonItems } from "./UseGetCreateButtonItems";
